+++ src/main/java/com/multi/metahouse/chat/controller/ChatController.java
@@ -1,4 +1,4 @@
-package com.multi.metahouse.chat;
+package com.multi.metahouse.chat.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
