+++ src/main/java/com/multi/metahouse/main/controller/indexcontroller.java
@@ -1,4 +1,4 @@
-package com.multi.metahouse.main;
+package com.multi.metahouse.main.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
