+++ src/main/java/com/multi/metahouse/member/controller/MemberController.java
@@ -1,4 +1,4 @@
-package com.multi.metahouse.member;
+package com.multi.metahouse.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
